- Refactoring of models to better packages - extending version model with type and description
