+++ src/main/java/com/repomgr/repomanager/infrastructure/VersionService.java
@@ -3,7 +3,11 @@ package com.repomgr.repomanager.infrastructure;
 import com.repomgr.repomanager.constants.Constants;
 import com.repomgr.repomanager.infrastructure.model.VersionEntity;
 import com.repomgr.repomanager.infrastructure.repository.VersionRepository;
-import com.repomgr.repomanager.rest.model.*;
+import com.repomgr.repomanager.rest.model.artifacts.VersionInformationContainerDto;
+import com.repomgr.repomanager.rest.model.artifacts.VersionInformationDto;
+import com.repomgr.repomanager.rest.model.common.MessageDto;
+import com.repomgr.repomanager.rest.model.common.PageDto;
+import com.repomgr.repomanager.rest.model.common.ResponseDto;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.BeanUtils;
