+++ src/main/resources/static/swagger.yaml
@@ -304,7 +304,7 @@ definitions:
         example: "MyLibrary"
 
   VersionInformation:
-    description: "Contains informations about the artifacts."
+    description: "Contains information about the artifacts."
     allOf:
       - $ref: '#/definitions/VersionInformationFilter'
       - type: object
@@ -315,6 +315,7 @@ definitions:
           - artifactId
           - version
           - creationDate
+          - type
         properties:
           repositoryUrl:
             description: "URL of the repository (optional)"
